added a message tester for game controller and combined all ssl-sockets into 1 file
