+++ src/TeamControl/network/__init__.py
@@ -1,15 +1,10 @@
 __all__ = ["BaseSocket","SocketType","Receiver","Broadcast","Multicast","Sender","Broadcaster","Multicaster",
-           "grSimYellowControl","grSimBlueControl","grSimControl","grSimSender","grSimVision","Vision","VisionTracker","GameControl",
+           "grSimSender","grSimVision","Vision","VisionTracker","GameControl",
            "GSRobotCommand","RobotCommand"]
 from TeamControl.network.baseUDP import BaseSocket,SocketType
 from TeamControl.network.receiver import Receiver, Broadcast, Multicast
 from TeamControl.network.sender import Sender,Broadcaster,Multicaster
 
-from TeamControl.network.visionSockets import Vision,VisionTracker
+from TeamControl.network.ssl_sockets import Vision,VisionTracker,grSimVision,grSimSender,GameControl
 
 from TeamControl.network.robotCommand import RobotCommand
-
-if __name__ == "__main__" : 
-    from TeamControl.network import sender
-    recv = sender()
-
